Fix syntax and remove redundant function
Replace the redundant code with list api in q_swap and q_shuffle to make
it more elegant.

Remove the swap_two_node function in queue.c and queue_expansion.h cause it won't be used again. Also temporary remove the web function cause it was not implemented in right way.